requirements/base.txt
Django==1.5.4
MySQL-python==1.2.4
South==0.8.2
Unipath==1.0
dj-database-url==0.2.2
django-braces==1.2.2
django-filter==0.7
django-model-utils==1.5.0
